Added new template functionality
